[Refactor]: inference 폴더 생성
- 기존 inference.py 가 하는 일이 너무 많고, 지저분함
- inference 폴더를 생성하여 역할을 구분
- 기존 inference.py는 init_model로 이름 변경
