+++ backend/src/AI/inference/__init__.py
@@ -1,0 +1,3 @@
+from .newmf_inference import InferenceNewMF
+from .mcn_inference import MCNInference
+from .simple_mcn_inference import SimpleMCNInference
